So, I've got a rather simple file which you give the corners and that has teh ability to roate these (using quaternions! :p)
Now, the aim is to convert this data to some surface, or polygon or whatever, so that it can be easily plotted. Then I can define a collection of these Face objects and plot D2d etc.

Also, it is currently *only* applicable to square surfaces. In the future, this might be extended. There are several factors on which that depends, the most important one if the group would appreciate it (because why waste time if they don't?)